add unit test & small changes
- Unit test (render, submit, enable/disable)
- Move to chat input to separate component
- Get focus on re-enable
