+++ frontend/package.json
@@ -21,6 +21,7 @@
     "@tailwindcss/postcss": "^4.1.13",
     "@testing-library/jest-dom": "^6.8.0",
     "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
     "@types/jest": "^29.5.8",
     "@types/node": "^20",
     "@types/react": "^18.3.24",
